Move classes into files. Documentation.
