+++ Source/Digst.OioIdws.OioWsTrust/Digst.OioIdws.OioWsTrust.csproj
@@ -57,6 +57,8 @@
     <Compile Include="ProtocolChannel\OioWsTrustChannel.cs" />
     <Compile Include="ProtocolChannel\OioWsTrustChannelFactory.cs" />
     <Compile Include="ProtocolChannel\OioWsTrustMessageTransformer.cs" />
+    <Compile Include="StsAuthenticationCase.cs" />
+    <Compile Include="StsTokenServiceBase.cs" />
     <Compile Include="StsTokenServiceConfiguration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReplyAttackCache.cs" />
